% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{decompress}
\alias{decompress}
\title{decompress files}
\usage{
decompress(file, outdir = NULL)
}
\arguments{
\item{file}{需要进行解压的压缩文件名字}

\item{outdir}{解压后的文件路径名}
}
\description{
目前支持zip, tar.gz, tar.bz2, tar, gz
}
