% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{get_color_palette}
\alias{get_color_palette}
\title{get color themes}
\usage{
get_color_palette(
  n,
  alpha = 1,
  theme = c("simple", "maftools_mutation1", "maftools_mutation2", "maftools_snv",
    "cellchat", "protigy")
)
}
\arguments{
\item{n}{the number of color}

\item{alpha}{the color transparency. range [0,1].}

\item{theme}{use color from which theme}
}
\value{
return color code
}
\description{
get color themes
}
