% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_diff.R
\name{group_diff}
\alias{group_diff}
\title{run differential expression analysis by group}
\usage{
group_diff(
  data,
  group,
  method = c("wilcox", "Ttest"),
  log2FC = 1,
  pval = 0.05,
  fdr = 0.05,
  part = c("all", "up", "down")
)
}
\arguments{
\item{data}{the data matrix to calculate the SD.
For the matrix, the rows represent the genomic features, and the columns represents the samples.}

\item{group}{A data frame contain two columns. The first column is sample name matched with colnames of data,
The second column is the cluster label of samples.}

\item{method}{The methods to calculate the differential features}

\item{log2FC}{The log2 fold change cutoff, default 1}

\item{pval}{The p value cutoff, default 0.05}

\item{fdr}{The FDR cutoff, default 0.05}

\item{part}{Select the unregulated features}
}
\description{
differential expression analysis by group
}
