% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize dat with different method}
\usage{
normalize_data(
  data_matrix,
  method = c("median", "mean", "quantile", "upper_quantile", "VSN", "2-component")
)
}
\arguments{
\item{data_matrix}{The data matrix with column in sample and row in feature}

\item{method}{different methods to normalize data.}
}
\value{

}
\description{
Normalize dat with different method
}
