% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot the data distribution of each sample}
\usage{
plot_distribution(dat_matrix, color = "#EF562D")
}
\arguments{
\item{dat_matrix}{The data matrix with column in sample and row in feature}

\item{color}{Character giving the colour of the plot}
}
\value{
Return a `ggplot` object
}
\description{
Plot the data distribution of each sample
}
