% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pca}
\alias{plot_pca}
\title{PCA analysis of data}
\usage{
plot_pca(data, group, pic_title = "All-PCA", ellipse = FALSE)
}
\arguments{
\item{data}{A matrix representing the genomic data such as gene expression data, miRNA expression data.\cr
For the matrix, the rows represent the genomic features, and the columns represent the samples.}

\item{group}{A data frame contain two columns. The first column is sample name matched with colnames of data,
The second column is the cluster label of samples.}

\item{pic_title}{The title of plot.}

\item{ellipse}{Whether add the confidence ellipse.}
}
\value{
Return a `ggplot` object contained a PCA plot.
}
\description{
PCA analysis of data
}
