% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_venn}
\alias{plot_venn}
\title{Draw Venn plot}
\usage{
plot_venn(x, filename = NULL, width = 7, height = 7, ...)
}
\arguments{
\item{x}{A list of vectors (e.g., integers, chars), with each component corresponding to a separate circle in the Venn diagram}

\item{filename}{Filename for image output, or if NULL returns the grid object itself}

\item{width}{Integer giving the width of the output figure in units}

\item{height}{Integer giving the height of the output figure in units}

\item{...}{other argument of \code{magrittr::\link[magrittr:pipe]{\%>\%}}}
}
\description{
Draw Venn plot
}
