% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{quantile_normalize}
\alias{quantile_normalize}
\title{Quantile normalization of data}
\usage{
quantile_normalize(data_matrix)
}
\arguments{
\item{data_matrix}{The data matrix with column in sample and row in feature}
}
\value{
return data matrix after normalize
}
\description{
Quantile normalization of data
}
