% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_all_sheets}
\alias{read_all_sheets}
\title{Read all sheets in an excel file}
\usage{
read_all_sheets(file, sheets = NULL, merge = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{file}{path of the .xls or .xlsx file}

\item{sheets}{sheet name or number to read, if NULL read all sheets}

\item{merge}{whether merge the data}

\item{...}{other arguments of `read_excel` function}

\item{verbos}{whether print useful message}
}
\description{
Read all sheets in an excel file
}
