% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TvsN.R
\name{tumor_vs_normal}
\alias{tumor_vs_normal}
\title{Tumor vs Normal
Adjust tumor expression by Normal sample expression}
\usage{
tumor_vs_normal(data, tumor = NULL, normal = NULL, log = FALSE)
}
\arguments{
\item{data}{A matrix representing the genomic data such as gene expression data, miRNA expression data.\cr
For the matrix, the rows represent the genomic features, and the columns represent the samples.}

\item{tumor}{A vector contain the names of tumor samples, it must match with `data` column names.}

\item{normal}{A vector contain the names of normal samples, it must match with `data` column names.
And with matched order with `tumor`.}

\item{log}{whether have log transform the `data`}
}
\description{
Tumor vs Normal
Adjust tumor expression by Normal sample expression
}
