% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{two_comp_normalize}
\alias{two_comp_normalize}
\title{Two component normalize of data}
\usage{
two_comp_normalize(sample, type = "default", mode.lower.bound = -3)
}
\arguments{
\item{sample}{the value of one sample}

\item{type}{types}

\item{mode.lower.bound}{mode.lower.bound}
}
\description{
Two component normalize of data
}
