% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{upper_quantile_normalize}
\alias{upper_quantile_normalize}
\title{Normalize data use upper quantile}
\usage{
upper_quantile_normalize(data_matrix)
}
\arguments{
\item{data_matrix}{The data matrix with column in sample and row in feature}
}
\value{
return data matrix after normalize
}
\description{
Normalize data use upper quantile
}
