% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{vsn_normalize}
\alias{vsn_normalize}
\title{VSN normalization of data}
\usage{
vsn_normalize(data_matrix)
}
\arguments{
\item{data_matrix}{The data matrix with column in sample and row in feature}
}
\value{
return data matrix after normalize
}
\description{
VSN normalization of data
}
